Created location class; uses JSON to load waypoint
A new `Location(latitude, longitude)` class now manages location
coordinate points more efficiently. When the main class starts,
location data is read in from `locations.json` in the main source
folder. The buoy data will be entered in the file during competition to
avoid interfacing with the code of the program.