Print output converted to basic logging framework
The framework for the logging system has been implemented. Using Python
native libraries, the application logs all output / feedback to the
`stdout console` and to a log file defined in the configuration if the
`debug` option is enabled. In future, this data will be transmitted to
the client through the web socket as well. I think the program’s
ability to give feedback in simple and yet all-encompassing ways is
highly important to the success of the project as a whole. Our ways of
gathering data in the future should not be limited to looking at the
command line where the application is being run. A permanent log and a
more graphic output system is needed.