All JSON now includes the `category` parameter
All JSON sent to the web socket server should include the `category`
tag. This is used to correctly distinguish the packet type on the
client side and process it correctly and more efficiently. This will
also help to support other web socket logging in the future. At present
the `data` tag is appended to all data which should be rendered in the
sidebar panel. The `log` tag is included in all messages that should be
rendered in the client terminal.