Implemented new GPSD Python TCP Socket
In order to leverage the native Python 2 GPSD libraries on the
Raspberry Pi, a daemon was created to run parallel to the core program.
It creates a TCP socket polls data from GPSD, creating a dictionary
object that is updated as data is continuously streamed in from GPSD.
When the client connects, it can send the character `0` to request the
object to be sent through the socket.